include criteria for analysis
The current model was crashing when using v8 files as there is no frame count column in it. Updated it so that the analysis now has an option to include the v8 files or not.
Further, the speed and the time doesn't have any values which are from v8 files in any case.