Refactor CSV writing and increase flush interval
Refactored the video processing loop to keep CSV files open for the duration of the segment, writing and flushing rows in batches instead of reopening files each time. Increased the default flush interval from 30 to 600 frames and added support for config override. This improves performance and reliability when writing large amounts of detection data.